requirements.txt
pytorch==1.13.1
torchvision==0.14.1
torchaudio==0.13.1
pytorch-cuda==11.7
transformers==4.46.3
numpy==1.24.4
pandas==2.0.3
scikit-learn==1.3.2
scipy==1.10.1
matplotlib==3.3.2
seaborn==0.13.0
biopython==1.83
umap-learn==0.5.6